Inference/requirements.txt
bitsandbytes==0.45.2
email_validator==2.2.0
fastapi-cli==0.0.7
fqdn==1.5.1
git-filter-repo==2.47.0
hf_transfer==0.1.9
httptools==0.6.4
inquirerpy==0.3.4
isoduration==20.11.0
jsonpointer==3.0.0
jupyter==1.1.1
liger_kernel==0.6.4
python-dotenv==1.1.0
python-multipart==0.0.20
scikit-learn==1.7.2
seaborn==0.13.2
tinycss2==1.4.0
torchao==0.11.0
trl==0.25.1
uri-template==1.3.0
uvloop==0.21.0
vllm==0.8.5.post1
webcolors==24.11.1
websockets==15.0.1
wheel==0.45.1